+++ src/plonemeeting/portal/core/content/subscriber.py
@@ -44,15 +44,15 @@ def handle_institution_modified(institution, event):
     """
     Initializes categories_mappings by trying to match global categories with fetched categories from iA.Delib.
     """
-    if hasattr(institution, "delib_categories") and not institution.categories_mappings:
+    local_categories = getattr(institution, institution.get_delib_categories_attr_name(), {})
+    if local_categories and not institution.categories_mappings:
         logger.info("Initializing default categories mappings by matching with those fetched from iA.Delib.")
         global_categories = [cat for cat in get_registry_record(name="plonemeeting.portal.core.global_categories")]
-        local_categories = [cat[0] for cat in institution.delib_categories]
         categories_mappings = []
-        for local_category in local_categories:
-            if local_category in global_categories:
-                categories_mappings.append({"local_category_id": local_category,
-                                            "global_category_id": local_category})
+        for local_category_id in local_categories.keys():
+            if local_category_id in global_categories:
+                categories_mappings.append({"local_category_id": local_category_id,
+                                            "global_category_id": local_category_id})
         institution.categories_mappings = categories_mappings
         logger.info("{} fetched iA.Delib categories matched.".format(len(categories_mappings)))