+++ CHANGES.rst
@@ -13,6 +13,8 @@ Changelog
   [odelaere]
 - Query representatives from iA.Delib to populate vocabularies only when loading the edit form.
   [odelaere]
+- Fetched data from delib are kept if used.
+  [odelaere]
 
 
 1.3.3.2 (2021-08-20)